# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScamShield AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageInput {
  /**
   * The text message to analyse for scam indicators
   * @minLength 1
   */
  message: string;
}

/**
 * Overall verdict category
 */
export type ScamAnalysisVerdict =
  (typeof ScamAnalysisVerdict)[keyof typeof ScamAnalysisVerdict];

export const ScamAnalysisVerdict = {
  safe: "safe",
  suspicious: "suspicious",
  likely_scam: "likely_scam",
  definite_scam: "definite_scam",
} as const;

export type ScamIndicatorType =
  (typeof ScamIndicatorType)[keyof typeof ScamIndicatorType];

export const ScamIndicatorType = {
  red_flag: "red_flag",
  yellow_flag: "yellow_flag",
  safe_signal: "safe_signal",
} as const;

export interface ScamIndicator {
  type: ScamIndicatorType;
  description: string;
}

export interface ScamAnalysis {
  /**
   * Likelihood of the message being a scam (0-100)
   * @minimum 0
   * @maximum 100
   */
  scamPercentage: number;
  /** Overall verdict category */
  verdict: ScamAnalysisVerdict;
  /** Plain-language explanation of the analysis */
  explanation: string;
  /** Specific red flags or reassuring factors found */
  indicators: ScamIndicator[];
}

export interface ErrorResponse {
  error: string;
}