# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScamShield AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scam
    description: Scam detection operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /scam/analyse:
    post:
      operationId: analyseMessage
      tags: [scam]
      summary: Analyse a message for scam indicators
      description: Uses AI to analyse text for phishing, scam, and social engineering indicators
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScamAnalysis"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    MessageInput:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
          description: The text message to analyse for scam indicators
    ScamAnalysis:
      type: object
      required:
        - scamPercentage
        - verdict
        - explanation
        - indicators
      properties:
        scamPercentage:
          type: integer
          minimum: 0
          maximum: 100
          description: Likelihood of the message being a scam (0-100)
        verdict:
          type: string
          enum: [safe, suspicious, likely_scam, definite_scam]
          description: Overall verdict category
        explanation:
          type: string
          description: Plain-language explanation of the analysis
        indicators:
          type: array
          items:
            $ref: "#/components/schemas/ScamIndicator"
          description: Specific red flags or reassuring factors found
    ScamIndicator:
      type: object
      required:
        - type
        - description
        - severity
      properties:
        type:
          type: string
          enum: [red_flag, yellow_flag, safe_signal]
        description:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string