# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScamShield AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  /**
   * The text message to analyse for scam indicators
   * @minLength 1
   */
  message: string;
}