# lib/api-zod/src/generated/types/scamAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScamShield AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScamAnalysisVerdict } from "./scamAnalysisVerdict";
import type { ScamIndicator } from "./scamIndicator";

export interface ScamAnalysis {
  /**
   * Likelihood of the message being a scam (0-100)
   * @minimum 0
   * @maximum 100
   */
  scamPercentage: number;
  /** Overall verdict category */
  verdict: ScamAnalysisVerdict;
  /** Plain-language explanation of the analysis */
  explanation: string;
  /** Specific red flags or reassuring factors found */
  indicators: ScamIndicator[];
}