# lib/api-zod/src/generated/types/scamIndicator.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScamShield AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScamIndicatorType } from "./scamIndicatorType";

export interface ScamIndicator {
  type: ScamIndicatorType;
  description: string;
}